+++ theme/src/utils/index.js
@@ -1,0 +1,12 @@
+export const soundCloudRegEx = /\/\/.[^(")|(<)]+/;
+
+export function isUrlValid(url) {
+  var res = url.match(
+    /(http(s)?:\/\/.)?(www\.)?[-a-zA-Z0-9@:%._+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_+.~#?&//=]*)/g
+  );
+  if (res != null) {
+    return true;
+  } else {
+    return false;
+  }
+}
